마이페이지도 회원에게만 접근가능하게 조정
마이페이지에서 수정하도록
- udpateMember메소드에서 passwordEncoder로 비밀번호 확인하고 다를 경우 js를 통해 에러 메세지 내보내도록

삭제나 수정 등 주소값을 만약 아무나 입력해도 적용이 안되게
- 멤버와 게시글 수정 삭제 주소에 {id}를 추가하고 이를 통해 인증된 사용자의 Id와 {id}의 값이 일치하는지 확인
- 확인 결과 일치하지 않으면 권한없음 페이지로 보냄

관리자 페이지, 오류페이지 제외하고 nav바 전부 동일하게 추가

+++ BoardWeb/src/main/java/com/fastwon/board/controller/BoardController.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -100,9 +101,12 @@ public class BoardController {
 	}
 
 	
-	@GetMapping("/updateBoard")
-	public String updateBoardView(Board board, Model model) {
-		Board findBoard = boardService.getUdateBoard(board);
+	@GetMapping("/{id}/updateBoard")
+	public String updateBoardView(@PathVariable String id, Board board, Model model, @AuthenticationPrincipal SecurityUser principal) {
+		if(!id.equals(principal.getMember().getId())) {
+			return "system/accessDenied";
+		}
+		Board findBoard = boardService.getUpdateBoard(board);
 		
 		model.addAttribute("board", findBoard);
 		return "board/updateBoard";
@@ -127,7 +131,7 @@ public class BoardController {
             //board 객체에 이미지 URL을 저장하는 필드를 추가
             board.setPhotoUrl("/uploads/" + fileName);
         } else {
-        	Board findBoard = boardService.getUdateBoard(board);
+        	Board findBoard = boardService.getUpdateBoard(board);
         	board.setPhotoUrl(findBoard.getPhotoUrl());
         }
 		
@@ -135,9 +139,12 @@ public class BoardController {
 		return "redirect:getBoardList";
 	}
 	
-	@GetMapping("deleteBoard")
-	public String deleteBoard(Board board) {
+	@GetMapping("/{id}/deleteBoard")
+	public String deleteBoard(@PathVariable String id, Board board, @AuthenticationPrincipal SecurityUser principal) {
+		if(!id.equals(principal.getMember().getId())) {
+			return "system/accessDenied";
+		}
 		boardService.deleteBoard(board);
-		return "redirect:getBoardList";
+		return "redirect:../getBoardList";
 	}
 }
